[6625] Make the project title more compact
Bug: https://github.com/eclipse-sirius/sirius-web/issues/6625
