+++ packages/sirius-web/frontend/sirius-web-application/src/views/edit-project/navbar/EditProjectNavbar.tsx
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2019, 2025 Obeo.
+ * Copyright (c) 2019, 2026 Obeo.
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v2.0
  * which accompanies this distribution, and is available at
@@ -10,46 +10,14 @@
  * Contributors:
  *     Obeo - initial API and implementation
  *******************************************************************************/
-import MoreVertIcon from '@mui/icons-material/MoreVert';
-import IconButton from '@mui/material/IconButton';
-import Typography from '@mui/material/Typography';
-import { emphasize } from '@mui/material/styles';
-import React, { useEffect, useState } from 'react';
-import { makeStyles } from 'tss-react/mui';
+import Box from '@mui/material/Box';
+import { useEffect, useState } from 'react';
 import { NavigationBar } from '../../../navigationBar/NavigationBar';
 import { useCurrentProject } from '../useCurrentProject';
 import { EditProjectNavbarProps, EditProjectNavbarState } from './EditProjectNavbar.types';
-import { EditProjectNavbarContextMenu } from './context-menu/EditProjectNavbarContextMenu';
+import { ProjectTitle } from './ProjectTitle';
 import { useProjectSubscription } from './useProjectSubscription';
 import { GQLProjectEventPayload, GQLProjectRenamedEventPayload } from './useProjectSubscription.types';
-
-const useEditProjectViewNavbarStyles = makeStyles()((theme) => ({
-  center: {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-  },
-  title: {
-    display: 'flex',
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  titleLabel: {
-    marginRight: theme.spacing(2),
-    overflow: 'hidden',
-    textOverflow: 'ellipsis',
-    whiteSpace: 'nowrap',
-    maxWidth: '100ch',
-  },
-  onDarkBackground: {
-    '&:hover': {
-      backgroundColor: emphasize(theme.palette.secondary.main, 0.08),
-    },
-  },
-  icon: {
-    fontSize: '1.25rem',
-  },
-}));
 
 const isProjectRenamedEventPayload = (payload: GQLProjectEventPayload): payload is GQLProjectRenamedEventPayload =>
   payload.__typename === 'ProjectRenamedEventPayload';
@@ -57,19 +25,15 @@ const isProjectRenamedEventPayload = (payload: GQLProjectEventPayload): payload 
 export const EditProjectNavbar = ({ workbenchHandle }: EditProjectNavbarProps) => {
   const { project } = useCurrentProject();
   const [state, setState] = useState<EditProjectNavbarState>({
-    anchorEl: null,
     projectName: project.name,
   });
 
   useEffect(() => {
     setState((prevState) => ({
       ...prevState,
-      anchorEl: null,
       projectName: project.name,
     }));
   }, [project]);
-
-  const { classes } = useEditProjectViewNavbarStyles();
 
   const { payload } = useProjectSubscription(project.id);
   useEffect(() => {
@@ -81,49 +45,17 @@ export const EditProjectNavbar = ({ workbenchHandle }: EditProjectNavbarProps) =
     }
   }, [payload]);
 
-  const onMoreClick = (event: React.MouseEvent<HTMLElement>) =>
-    setState((prevState) => ({
-      ...prevState,
-      anchorEl: event.currentTarget,
-    }));
-
-  const onCloseContextMenu = () =>
-    setState((prevState) => ({
-      ...prevState,
-      anchorEl: null,
-    }));
-
   return (
-    <>
-      <NavigationBar>
-        <div className={classes.center}>
-          <div className={classes.title}>
-            <Typography variant="h6" noWrap className={classes.titleLabel} data-testid={`navbar-title`}>
-              {state.projectName}
-            </Typography>
-            <IconButton
-              className={classes.onDarkBackground}
-              edge="start"
-              size="small" // Per #3591 it should remain "small" to keep vertical space for a potential subtitle
-              aria-label="more"
-              aria-controls="more-menu"
-              aria-haspopup="true"
-              onClick={onMoreClick}
-              color="inherit"
-              data-testid="more">
-              <MoreVertIcon className={classes.icon} />
-            </IconButton>
-          </div>
-        </div>
-      </NavigationBar>
-      {state.anchorEl ? (
-        <EditProjectNavbarContextMenu
-          anchorEl={state.anchorEl}
-          projectName={state.projectName}
-          onClose={onCloseContextMenu}
-          workbenchHandle={workbenchHandle}
-        />
-      ) : null}
-    </>
+    <NavigationBar>
+      <Box
+        sx={{
+          display: 'grid',
+          gridTemplateRows: '1fr',
+          gridTemplateColumns: '1fr',
+          alignItems: 'center',
+        }}>
+        <ProjectTitle name={state.projectName} workbenchHandle={workbenchHandle} key={project.id} />
+      </Box>
+    </NavigationBar>
   );
 };
